Move update essentials from enkf_main to analysis/update
